+++ src/interfaces/index.ts
@@ -1,3 +1,0 @@
-export interface HandleChange {
-  target: HTMLInputElement;
-}