+++ src/types/index.ts
@@ -1,41 +1,58 @@
 import { API } from '@storybook/api';
+import { ReactNode } from 'react';
 
 /*
   Theme
 */
 export type ThemeObject = { name: string; theme: object };
-export type ThemesArray = Array<ThemeObject>;
-export type Theme = ThemesArray | { [key: string]: any };
+export type ThemesArray = ThemeObject[];
+
+export type LabelFormatFunction = (path: string[]) => string;
 
 /*
   Config
 */
 export type ConfigProps = {
-  labelFormat: any;
-  showCode: boolean;
-  debounce: boolean;
-  debounceRate: number;
+  labelFormat?: 'startCase' | 'path' | LabelFormatFunction;
+  showCode?: boolean;
+  debounce?: boolean;
+  debounceRate?: number;
 };
 
 /*
   Options
 */
 export type OptionsType = {
-  theme: Theme;
-  overrides: Overrides;
+  theme: any;
+  overrides: OverridesProps;
   config: ConfigProps;
 };
 
 /*
   Overrides
 */
-export type OverridesObject = {
-  type: string;
-  [key: string]: any;
-};
+export type OverridesProps = {
+  [key: string]: {
+    type:
+      | 'color'
+      | 'counter'
+      | 'select'
+      | 'shorthand'
+      | 'switch'
+      | 'radio'
+      | 'range';
 
-export type Overrides = {
-  [key: string]: OverridesObject;
+    hidden?: boolean;
+    label?: string;
+    description?: string;
+    min?: number;
+    max?: number;
+    steps?: number;
+    options?: {
+      value: string | number;
+      label: string;
+    }[];
+  };
 };
 
 /*
@@ -43,16 +60,17 @@ export type Overrides = {
 */
 export type SettingsProviderProps = {
   api: API;
+  children: ReactNode;
 };
 
 export type SettingsContextProps = {
   themes: ThemesArray;
   activeTheme: ThemeObject;
-  themeComponents: {};
-  overrides: Overrides;
+  themeComponents: any;
+  overrides: OverridesProps;
   config: ConfigProps;
   isLoading: boolean;
-  updateTheme: (path: any, value: any) => void;
+  updateTheme: (path: string, value: any) => void;
   updateActiveTheme: (obj: ThemeObject) => void;
   resetThemes: () => void;
 };