+++ CHANGELOG.md
@@ -7,11 +7,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.3.4]- 2020-09-06
+
+### Added
+
+- [Issue #14](https://github.com/jeslage/storybook-addon-theme-playground/issues/14): Exported `ThemePlaygroundProps` interface
+
+### Fixed
+
+- [Issue #14](https://github.com/jeslage/storybook-addon-theme-playground/issues/14): Optional properties `config` and `overrides` where required inside ThemePlayground `options`
+
 ## [1.3.3]- 2020-06-27
 
 ### Changed
 
-- [Issue #3](https://github.com/jeslage/storybook-addon-theme-playground/issues/3): Moved `withThemePlayground` to a class component, due to invalid hook call while using `@storybook/addon-storyshots`.
+- [Issue #3](https://github.com/jeslage/storybook-addon-theme-playground/issues/3): Moved `withThemePlayground` to a class component due to invalid hook call while using `@storybook/addon-storyshots`.
 
 ## [1.3.2]- 2020-06-27
 
@@ -100,7 +110,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release
 
-[unreleased]: https://github.com/jeslage/storybook-addon-theme-playground/compare/v1.3.3...develop
+[unreleased]: https://github.com/jeslage/storybook-addon-theme-playground/compare/v1.3.4...develop
+[1.3.4]: https://github.com/jeslage/storybook-addon-theme-playground/compare/v1.3.4...1.3.3
 [1.3.3]: https://github.com/jeslage/storybook-addon-theme-playground/compare/v1.3.3...1.3.2
 [1.3.2]: https://github.com/jeslage/storybook-addon-theme-playground/compare/v1.3.1...1.3.2
 [1.3.1]: https://github.com/jeslage/storybook-addon-theme-playground/compare/v1.3.0...1.3.1